fix: Rename to PermissionsProvider
