+++ src/routes/issues.tsx
@@ -3,7 +3,7 @@ import useActiveRedmineTab from "@/hooks/useActiveRedmineTab";
 import { issuePrioritiesQueryOptions, useSuspenseIssuePriorities } from "@/hooks/useIssuePriorities";
 import { myOpenIssuesQueryOptions, useSuspenseMyIssues } from "@/hooks/useMyIssues";
 import { useSuspenseMultipleProjectVersions } from "@/hooks/useProjectVersions";
-import PermissionProvider from "@/provider/PermissionProvider";
+import PermissionProvider from "@/provider/PermissionsProvider";
 import { useSettings } from "@/provider/SettingsProvider";
 import { groupIssues } from "@/utils/groupIssues";
 import { createFileRoute } from "@tanstack/react-router";
